cast sizeof() to int for format specifier
sizeof() returns size_t, whereas the format string expects %d (an int).
Cast the result of sizeof() to an int to satisfy the format string.

It's possible that this was done to fix a compiler warning.

(Another way to handle this would be to use the %zu format specifier
instead of %d.)

Signed-off-by: Ken Dreyer <kdreyer@redhat.com>